+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 Upcoming
 +++++++
+* 'az containerapp env update': Fix bugs in update environment.
 * Fix YAML create with user-assigned identity
 * Fix polling logic for long running operations.